scripts/requirements.txt
# For scripts/urdf_to_cad_params.py
PyYAML>=5.1
